Extract dumper/loader and conversion logic
This commit moves conversion functions, dumpers, and loaders into
respective separate files.  With this update, maintenance of the adapter
will be easier to reason about.

* Loading and Dumping functions will be isolated and easier to test
* Conversion functions can be made available to additional modules
